moved Q3D to glider
